creates some pakcages, sets up some stubs
