+++ src/main/java/csci331/team/red/shared/Person.java
@@ -1,0 +1,5 @@
+package csci331.team.red.shared;
+
+public class Person {
+
+}
